+++ GrowIT/Feat/Mypage/Presentation/ViewControllers/MyAccountViewController.swift
@@ -261,18 +261,7 @@ extension MyAccountViewController: UITableViewDataSource, UITableViewDelegate {
                
                 let modalVC = CannotChangePasswordViewController()
                 modalVC.modalPresentationStyle = .pageSheet
-                if let sheet = modalVC.sheetPresentationController {
-                    //지원할 크기 지정
-                    if #available(iOS 16.0, *) {
-                        sheet.detents = [
-                            .custom{ context in
-                                0.36 * context.maximumDetentValue
-                            }
-                        ]
-                    } else { sheet.detents = [.medium()] }
-                    sheet.prefersGrabberVisible = true
-                }
-                present(modalVC, animated: true, completion: nil)
+                presentSheet(modalVC, heightRatio: 0.36)
             }
         case (1, 0):
             didTapTermsOfService("개인정보 수집•이용 동의")
